+++ routes/api.php
@@ -6,20 +6,19 @@ use App\Http\Controllers\Administrator\NoticeController;
 use App\Http\Controllers\AfterService\AfterServiceCommentController;
 use App\Http\Controllers\AfterService\AfterServiceController;
 use App\Http\Controllers\Auth\AdminController;
+use App\Http\Controllers\Auth\PasswordResetCodeController;
+use App\Http\Controllers\Auth\RefreshController;
 use App\Http\Controllers\Auth\UserController;
 use App\Http\Controllers\MeetingRoom\MeetingRoomController;
 use App\Http\Controllers\MeetingRoom\MeetingRoomReservationController;
-use App\Http\Controllers\PasswordResetCodeController;
 use App\Http\Controllers\QRController;
-use App\Http\Controllers\RefreshController;
+use App\Http\Controllers\Restaurant\RestaurantAccountController;
+use App\Http\Controllers\Restaurant\RestaurantApplyDivisionController;
 use App\Http\Controllers\Restaurant\RestaurantMenusController;
 use App\Http\Controllers\Restaurant\RestaurantSemesterController;
 use App\Http\Controllers\Restaurant\RestaurantWeekendController;
 use App\Http\Controllers\Restaurant\SemesterMealTypeController;
 use App\Http\Controllers\Restaurant\WeekendMealTypeController;
-
-use App\Http\Controllers\Restaurant\RestaurantApplyDivisionController;
-use App\Http\Controllers\Restaurant\RestaurantAccountController;
 use App\Http\Controllers\Salon\SalonBreakTimeController;
 use App\Http\Controllers\Salon\SalonBusinessHourController;
 use App\Http\Controllers\Salon\SalonCategoryController;
